secure-chat/requirements.txt
cryptography
pynacl
websockets
rich
prompt_toolkit
